+++ docs/CHANGES.md
@@ -6,6 +6,7 @@ dev
  * MP3 failure returns with non-zero RC when only one input file is provided
  * Improved track title generation from file name
  * Moved usage documentation to separate document
+ * Improved Python3 code compatibility
 
 v1.1.0 (2018-10-04)
 -------------------